--- data/info_by_zipcode/descriptions.txt
@@ -1,1 +1,0 @@
-{'HC01_EST_VC01': 'Households; Estimate; Total', 'HC01_EST_VC02': 'Households; Estimate; Less than $10,000', 'HC01_EST_VC03': 'Households; Estimate; $10,000 to $14,999', 'HC01_EST_VC04': 'Households; Estimate; $15,000 to $24,999', 'HC01_EST_VC05': 'Households; Estimate; $25,000 to $34,999', 'HC01_EST_VC06': 'Households; Estimate; $35,000 to $49,999', 'HC01_EST_VC07': 'Households; Estimate; $50,000 to $74,999', 'HC01_EST_VC08': 'Households; Estimate; $75,000 to $99,999', 'HC01_EST_VC09': 'Households; Estimate; $100,000 to $149,999', 'HC01_EST_VC10': 'Households; Estimate; $150,000 to $199,999', 'HC01_EST_VC11': 'Households; Estimate; $200,000 or more', 'HC01_EST_VC13': 'Households; Estimate; Median income (dollars)', 'HC01_EST_VC15': 'Households; Estimate; Mean income (dollars)', 'HC01_EST_VC18': 'Households; Estimate; PERCENT IMPUTED - Household income in the past 12 months'}
