+++ data/shapefiles-us_zips/cb_2017_us_zcta510_500k.cpg
@@ -1,0 +1,1 @@
+UTF-8
